fix: Delete when behavior
