+++ packages/nodes-base/nodes/RabbitMQ/RabbitMQTrigger.node.ts
@@ -1,19 +1,16 @@
 /* eslint-disable n8n-nodes-base/node-filename-against-convention */
 import type { Message } from 'amqplib';
 import type {
-	IDeferredPromise,
-	IExecuteResponsePromiseData,
 	INodeProperties,
 	INodeType,
 	INodeTypeDescription,
-	IRun,
 	ITriggerFunctions,
 	ITriggerResponse,
 } from 'n8n-workflow';
 import { NodeConnectionTypes, NodeOperationError } from 'n8n-workflow';
 
 import { rabbitDefaultOptions } from './DefaultOptions';
-import { MessageTracker, rabbitmqConnectQueue, parseMessage } from './GenericFunctions';
+import { MessageTracker, rabbitmqConnectQueue, handleMessage } from './GenericFunctions';
 import type { TriggerOptions } from './types';
 
 export class RabbitMQTrigger implements INodeType {
@@ -207,131 +204,10 @@ export class RabbitMQTrigger implements INodeType {
 		const options = this.getNodeParameter('options', {}) as TriggerOptions;
 		const channel = await rabbitmqConnectQueue.call(this, queue, options);
 
-		if (this.getMode() === 'manual') {
-			const manualTriggerFunction = async () => {
-				// Do only catch a single message when executing manually, else messages will leak
-				await channel.prefetch(1);
-
-				const processMessage = async (message: Message | null) => {
-					if (message !== null) {
-						const item = await parseMessage(message, options, this.helpers);
-						channel.ack(message);
-						this.emit([[item]]);
-					} else {
-						this.emitError(new Error('Connection got closed unexpectedly'));
-					}
-				};
-
-				const existingMessage = await channel.get(queue);
-				if (existingMessage) await processMessage(existingMessage);
-				else await channel.consume(queue, processMessage);
-			};
-
-			const closeFunction = async () => {
-				await channel.close();
-				await channel.connection.close();
-				return;
-			};
-
-			return {
-				closeFunction,
-				manualTriggerFunction,
-			};
-		}
-
-		const parallelMessages = options.parallelMessages ?? -1;
-		if (isNaN(parallelMessages) || parallelMessages === 0 || parallelMessages < -1) {
-			throw new NodeOperationError(
-				this.getNode(),
-				'Parallel message processing limit must be a number greater than zero (or -1 for no limit)',
-			);
-		}
-
-		let acknowledgeMode = options.acknowledge ?? 'immediately';
-
-		if (parallelMessages !== -1 && acknowledgeMode === 'immediately') {
-			// If parallel message limit is set, then the default mode is "executionFinishes"
-			// unless acknowledgeMode got set specifically. Be aware that the mode "immediately"
-			// can not be supported in this case.
-			acknowledgeMode = 'executionFinishes';
-		}
-
 		const messageTracker = new MessageTracker();
+		let acknowledgeMode = options.acknowledge ?? 'immediately';
 		let closeGotCalled = false;
-
-		if (parallelMessages !== -1) {
-			await channel.prefetch(parallelMessages);
-		}
-
-		channel.on('close', () => {
-			if (!closeGotCalled) {
-				this.emitError(new Error('Connection got closed unexpectedly'));
-			}
-		});
-
-		const consumerInfo = await channel.consume(queue, async (message) => {
-			if (message !== null) {
-				try {
-					if (acknowledgeMode !== 'immediately') {
-						messageTracker.received(message);
-					}
-
-					const item = await parseMessage(message, options, this.helpers);
-
-					let responsePromise: IDeferredPromise<IRun> | undefined = undefined;
-					let responsePromiseHook: IDeferredPromise<IExecuteResponsePromiseData> | undefined =
-						undefined;
-					if (acknowledgeMode !== 'immediately' && acknowledgeMode !== 'laterMessageNode') {
-						responsePromise = this.helpers.createDeferredPromise();
-					} else if (acknowledgeMode === 'laterMessageNode') {
-						responsePromiseHook = this.helpers.createDeferredPromise<IExecuteResponsePromiseData>();
-					}
-					if (responsePromiseHook) {
-						this.emit([[item]], responsePromiseHook, undefined);
-					} else {
-						this.emit([[item]], undefined, responsePromise);
-					}
-					if (responsePromise && acknowledgeMode !== 'laterMessageNode') {
-						// Acknowledge message after the execution finished
-						await responsePromise.promise.then(async (data: IRun) => {
-							if (data.data.resultData.error) {
-								// The execution did fail
-								if (acknowledgeMode === 'executionFinishesSuccessfully') {
-									channel.nack(message);
-									messageTracker.answered(message);
-									return;
-								}
-							}
-							channel.ack(message);
-							messageTracker.answered(message);
-						});
-					} else if (responsePromiseHook && acknowledgeMode === 'laterMessageNode') {
-						await responsePromiseHook.promise.then(() => {
-							channel.ack(message);
-							messageTracker.answered(message);
-						});
-					} else {
-						// Acknowledge message directly
-						channel.ack(message);
-					}
-				} catch (error) {
-					const workflow = this.getWorkflow();
-					const node = this.getNode();
-					if (acknowledgeMode !== 'immediately') {
-						messageTracker.answered(message);
-					}
-
-					this.logger.error(
-						`There was a problem with the RabbitMQ Trigger node "${node.name}" in workflow "${workflow.id}": "${error.message}"`,
-						{
-							node: node.name,
-							workflowId: workflow.id,
-						},
-					);
-				}
-			}
-		});
-		const consumerTag = consumerInfo.consumerTag;
+		let consumerTag: string | undefined;
 
 		// The "closeFunction" function gets called by n8n whenever
 		// the workflow gets deactivated and can so clean up.
@@ -351,6 +227,66 @@ export class RabbitMQTrigger implements INodeType {
 				);
 			}
 		};
+
+		if (this.getMode() === 'manual') {
+			const manualTriggerFunction = async () => {
+				// Do only catch a single message when executing manually, else messages will leak
+				await channel.prefetch(1);
+
+				const processMessage = async (message: Message | null) => {
+					if (message !== null) {
+						handleMessage.call(this, message, channel, messageTracker, acknowledgeMode, options);
+					} else {
+						this.emitError(new Error('Connection got closed unexpectedly'));
+					}
+				};
+
+				const existingMessage = await channel.get(queue);
+				if (existingMessage) {
+					await processMessage(existingMessage);
+				} else {
+					const consumerInfo = await channel.consume(queue, processMessage);
+					consumerTag = consumerInfo.consumerTag;
+				}
+			};
+
+			return {
+				closeFunction,
+				manualTriggerFunction,
+			};
+		}
+
+		const parallelMessages = options.parallelMessages ?? -1;
+		if (isNaN(parallelMessages) || parallelMessages === 0 || parallelMessages < -1) {
+			throw new NodeOperationError(
+				this.getNode(),
+				'Parallel message processing limit must be a number greater than zero (or -1 for no limit)',
+			);
+		}
+
+		if (parallelMessages !== -1 && acknowledgeMode === 'immediately') {
+			// If parallel message limit is set, then the default mode is "executionFinishes"
+			// unless acknowledgeMode got set specifically. Be aware that the mode "immediately"
+			// can not be supported in this case.
+			acknowledgeMode = 'executionFinishes';
+		}
+
+		if (parallelMessages !== -1) {
+			await channel.prefetch(parallelMessages);
+		}
+
+		channel.on('close', () => {
+			if (!closeGotCalled) {
+				this.emitError(new Error('Connection got closed unexpectedly'));
+			}
+		});
+
+		const consumerInfo = await channel.consume(queue, async (message) => {
+			if (message !== null) {
+				handleMessage.call(this, message, channel, messageTracker, acknowledgeMode, options);
+			}
+		});
+		consumerTag = consumerInfo.consumerTag;
 
 		return {
 			closeFunction,
